Remove activitiName and activityType fields + split tests
