+++ activiti-spring-boot-starter/src/test/java/org/activiti/spring/boot/process/listener/DeployedProcessesListener.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
+ * Copyright 2019 Alfresco, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.activiti.spring.boot.process;
+package org.activiti.spring.boot.process.listener;
 
 import java.util.ArrayList;
 import java.util.HashMap;
